requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
python-socketio==5.15.0
aiofiles==25.1.0
python-multipart==0.0.20
pydantic-settings==2.7.1
